+++ R/output_format.R
@@ -21,7 +21,9 @@
 #'   for formats that produce HTML). In addition to the named methods you can
 #'   also pass an arbitrary function to be used for printing data frames. You
 #'   can disable the \code{df_print} behavior entirely by setting the option
-#'   \code{rmarkdown.df_print} to \code{FALSE}.
+#'   \code{rmarkdown.df_print} to \code{FALSE}. See
+#'   \href{https://bookdown.org/yihui/rmarkdown/html-document.html#data-frame-printing}{Data
+#'   frame printing section} in bookdown book for examples.
 #' @param pre_knit An optional function that runs before knitting which receives
 #'   the \code{input} (input filename passed to \code{render}) and \code{...}
 #'   (for future expansion) arguments.
@@ -260,7 +262,7 @@ knitr_options_pdf <- function(fig_width,
   knit_hooks <- NULL
 
   # apply cropping if requested and we have pdfcrop and ghostscript
-  crop <- fig_crop && find_program("pdfcrop") != '' && tools::find_gs_cmd() != ''
+  crop <- fig_crop && has_crop_tools()
   if (crop) {
     knit_hooks = list(crop = knitr::hook_pdfcrop)
     opts_chunk$crop = TRUE
@@ -772,12 +774,19 @@ is_pandoc_to_html <- function(options) {
 
 citeproc_required <- function(yaml_front_matter,
                               input_lines = NULL) {
+  # TODO: remove the hack below after BETS is updated on CRAN https://github.com/nmecsys/BETS/pull/18
+  if (tryCatch(xfun::check_old_package('BETS', '0.4.9'), error = function(e) FALSE)) return(FALSE)
   (
     is.null(yaml_front_matter$citeproc) ||
       yaml_front_matter$citeproc
   ) && (
     !is.null(yaml_front_matter$bibliography) ||
       !is.null(yaml_front_matter$references) ||
-      length(grep("^references\\:\\s*$", input_lines)) > 0
+      # detect references: and bibliography: outside of yaml header
+      # as Pandoc is supporting
+      # TODO: remove when supporting multiple yaml block
+      # https://github.com/rstudio/rmarkdown/issues/1891
+      length(grep("^references:\\s*$", input_lines)) > 0 ||
+      length(grep("^bibliography:\\s*$", input_lines)) > 0
   )
 }