+++ R/render_site.R
@@ -29,8 +29,9 @@
 #'       \item{Files beginning with "." (hidden files).}
 #'       \item{Files beginning with "_"}
 #'       \item{Files known to contain R source code (e.g. ".R", ".s", ".Rmd"), R
-#'       data (e.g. ".RData", ".rds"), or configuration data (e.g. ".Rproj",
-#'       "rsconnect")).}
+#'       data (e.g. ".RData", ".rds"), configuration data (e.g. ".Rproj",
+#'       "rsconnect") or package project management data (e.g.
+#'       "packrat", "renv").}
 #'     }
 #'     Note that you can override which files are included or excluded via
 #'     settings in "_site.yml" (described below).}
@@ -204,9 +205,9 @@ render_site <- function(input = ".",
 
 #' @rdname render_site
 #' @param preview Whether to list the files to be removed rather than actually
-#'   removing them.
+#'   removing them. Defaulting to TRUE to prevent removing without notice.
 #' @export
-clean_site <- function(input = ".", preview = FALSE, quiet = FALSE,
+clean_site <- function(input = ".", preview = TRUE, quiet = FALSE,
                        encoding = "UTF-8") {
 
   # normalize to a directory
@@ -220,17 +221,34 @@ clean_site <- function(input = ".", preview = FALSE, quiet = FALSE,
   # get the files to be cleaned
   files <- generator$clean()
 
+
+  if (length(files) == 0) {
+    if (preview || !quiet) cat("Nothing to removed. All clean !\n")
+    return(invisible(NULL))
+  }
+
   # if it's just a preview then return the files, otherwise
   # actually remove the files
-  if (preview)
-    files
-  else {
+  if (preview) {
+    cat("These files and folders can probably be removed:\n",
+        paste0("* ", mark_dirs(files)),
+        "\nUse rmarkdown::clean_site(preview = FALSE) to remove them.",
+        sep = "\n")
+  } else {
     if (!quiet) {
-      cat("Removing files: \n")
-      cat(paste0(" ", files), sep = "\n")
+      cat("Removing files: \n",
+          paste0("* ", mark_dirs(files)),
+          sep = "\n")
     }
     unlink(file.path(input, files), recursive = TRUE)
   }
+}
+
+# TODO: Move to xfun - bookdown use it too.
+mark_dirs <- function(files) {
+  i <- file_test("-d", files) & !grepl("/$", files)
+  files[i] <- paste0(files[i], "/")
+  files
 }
 
 #' @rdname render_site
@@ -320,9 +338,12 @@ site_config <- function(input = ".", encoding = "UTF-8") {
 # default site implementation (can be overridden by custom site generators)
 
 #' @rdname render_site
+#' @param output_format_filter An optional function which is passed the
+#'  input file and the output format, and which returns a (potentially
+#'  modified) output format.
 #' @param ... Currently unused.
 #' @export
-default_site_generator <- default_site <- function(input, ...) {
+default_site_generator <- default_site <- function(input, output_format_filter = NULL, ...) {
 
   # get the site config
   config <- site_config(input)
@@ -378,8 +399,14 @@ default_site_generator <- default_site <- function(input, ...) {
       # log the file being rendered
       if (!quiet) message("\nRendering: ", x)
 
+      # optionally customize the output format via filter
+      file_output_format <- output_format
+      if (is.function(output_format_filter)) {
+        file_output_format <- output_format_filter(x, output_format)
+      }
+
       output <- render_one(input = x,
-                           output_format = output_format,
+                           output_format = file_output_format,
                            output_options = list(lib_dir = "site_libs",
                                                  self_contained = FALSE),
                            envir = envir,
@@ -558,14 +585,14 @@ site_resources <- function(site_dir, include = NULL, exclude = NULL, recursive =
   # excludes:
   #   - known source/data extensions
   #   - anything that starts w/ '.' or '_'
-  #   - rsconnect and packrat directories
+  #   - rsconnect, renv and packrat directories
   #   - user excludes
   extensions <- c("R", "r", "S", "s",
                   "Rmd", "rmd", "md", "Rmarkdown", "rmarkdown",
                   "Rproj", "rproj",
                   "RData", "rdata", "rds")
   extensions_regex <- utils::glob2rx(paste0("*.", extensions))
-  excludes <- c("^rsconnect$", "^packrat$", "^\\..*$", "^_.*$", "^.*_cache$",
+  excludes <- c("^rsconnect$", "^packrat$", "^renv$", "^\\..*$", "^_.*$", "^.*_cache$",
                 extensions_regex,
                 utils::glob2rx(exclude))
   files <- all_files