+++ R/shiny_prerendered.R
@@ -4,7 +4,7 @@ shiny_prerendered_app <- function(input_rmd, render_args) {
 
   # get rendered html and capture dependencies
   html <- shiny_prerendered_html(input_rmd, render_args)
-  deps <- attr(html, "html_dependencies")
+  deps <- htmltools::htmlDependencies(html)
 
   # create the server environment
   server_envir = new.env(parent = globalenv())
@@ -691,4 +691,19 @@ shiny_prerendered_data_chunks_index <- function(data_dir) {
 shiny_prerendered_data_file_name <- function(label, cache) {
   type <- ifelse(cache, ".cached", "")
   sprintf("%s%s.RData", label, type)
+}
+
+# Use me instead of html_dependency_bootstrap() in a shiny runtime to get
+# dynamic theming (i.e., have it work with session$setCurrentTheme())
+shiny_bootstrap_lib <- function(theme) {
+  theme <- resolve_theme(theme)
+  if (!is_bs_theme(theme)) {
+    return(NULL)
+  }
+  if (!is_available("shiny", "1.6.0")) {
+    stop(
+      "Using a {bslib} theme with `runtime: shiny` requires shiny 1.6.0 or higher."
+    )
+  }
+  shiny::bootstrapLib(theme)
 }