+++ tests/testthat/test-lua-filters.R
@@ -1,14 +1,12 @@
 .generate_md_and_convert <- function(content, output_format) {
-  input_file <- tempfile(fileext = ".Rmd")
-  output_file <- tempfile()
-  on.exit(unlink(c(input_file, output_file)), add = TRUE)
-  xfun::write_utf8(c("---\ntitle: Test\n---\n", content), input_file)
+  input_file <- local_rmd_file(c("---\ntitle: Test\n---\n", content))
+  output_file <- withr::local_tempfile()
   res <- rmarkdown::render(input_file, output_format = output_format, output_file = output_file, quiet = TRUE)
   xfun::read_utf8(res)
 }
 
 # Lua filters exists only since pandoc 2.0
-skip_if_not(rmarkdown::pandoc_available("2.0"))
+skip_if_not_pandoc("2.0")
 
 test_that("pagebreak Lua filters works", {
   rmd <- "# HEADER 1\n\\newpage\n# HEADER 2\n\\pagebreak\n# HEADER 3"
@@ -32,6 +30,8 @@ test_that("number_sections Lua filter works", {
   rmd <- paste0(headers, " ", numbers, "\n\n")
   result <- .generate_md_and_convert(rmd, md_document(number_sections = TRUE))
   expected <- paste(numbers, numbers)
+  # pandoc 2.11.2 default to atx headers
+  if (pandoc_available("2.11.2")) expected <- paste(headers, expected)
   expect_identical(result[result %in% expected], expected)
 })